[変更]CommentCollectionクラスを別ファイルに移動
